Description
     Visually impaired people are the people who can't identify smallest detail 
with healthy eyes. Those who have the visual acuity of 6/60 or the horizontal extent 
of the visual field with both eyes open less than or equal to 20 degrees, these people 
are considered blind. Such people are in need of aiding devices for blindness related 
disabilities. As described in 10% of blind have no usable eyesight at all to help them 
move around independently and safely. The electronic aiding devices are designed 
to solve such issue.
     The Mobility Cane comes as a proposed solution to help visually impaired people to detect obstacles and investigate dangers in the world around while walking 
and also to cross the traffic signal without any accident. The system is designed to 
act like an artificial vision and alarm unit. 

      It consists of ultrasonic, LED, TCS 230Colour detecting sensor and ISD1820 
Voice recorder module in addition to the microcontroller (Arduino Uno) to receive 
the sensors signals and process them to short pulses to the Arduino pins where 
buzzer, TCS 230, ISD1820 Voice recorder module and LED are connected.